Copy reference files from correspnding newlib variants
libgcc and libgloss files are not built specifically for small/tiny
build variants but can be taken from newlib builds.  The new rules
calculate what the corresponding newlib build is called and take
an appropriate set of files.

As part of this the reliance on MIPS_INST_ROOT is removed as it
was not restricting the build to only use a compiler from that
location. The compiler is now assumed to be in the path and a
normal cmake error will occur if it is missing.